fixed occlusion shader only on one eye, new table spawn mechanics
Table Ghost is now spawned by pressing 'A'. It now follows the user's head and position and its position can be confirmed by pressing 'A' again.
To cancel the table's position placement, the ghost can be destroyed by pressing 'B'. (The issue of the markers not being at the right position is still being fixed under another branch)